feat: metadata for the pretty record is optional
The advantage of this approach is that the current data can still be rendered successfully even if the metadata is not updated in a timely manner. However, when fetching the latest data from GI, since the official data does not have an item_id, the fetch will still fail if the metadata is missing, which is still the expected behavior.